services/gee/requirements.txt
earthengine-api==1.4.2
google-cloud-bigquery==3.27.0
